requirements.txt
# Core dependencies
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# SAE and interpretability
sae-lens==6.12.1

# Data processing
datasets>=2.16.0
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=14.0.0

# ML utilities
scikit-learn>=1.3.0
scipy>=1.11.0

# Conformal prediction
mapie>=0.8.0

# CLI and config
fire>=0.5.0
pyyaml>=6.0
omegaconf>=2.3.0

# Logging and tracking
wandb>=0.16.0
tensorboard>=2.15.0
tqdm>=4.66.0

# Evaluation
rouge-score>=0.1.2
sacrebleu>=2.3.0

# Development
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# Optional: vLLM for faster inference
# vllm>=0.2.0  # Uncomment if using vLLM

